Redesign code to use configmap as the input
To be more practical, and ease the user experience
changed the way to configure the nad, by switching
from configuring nad using kpt kubeconfig, to using
configmaps to specify the nad configurations.

Done that by changing the tartget resource from being
nads, to be configmaps with a specific annotation. the
scripts would read the configmap, and create the nad
from the configmap data.

Signed-off-by: Abdallah Yasin <alsaffa71994@gmail.com>